requirements.txt
# ================================
# AutoGen 知识图谱项目核心依赖
# ================================

# ✨ 核心技术栈 (Core Tech Stack)
# ================================

# 🤖 Microsoft AutoGen - 多智能体框架
pyautogen==0.2.35

# 📊 Neo4j - 图数据库驱动
neo4j==5.15.0

# 🧠 OpenAI API - AI模型接口
openai==1.54.3

# 🚀 FastAPI - 现代Web框架
fastapi==0.109.0
uvicorn[standard]==0.27.0

# 📋 Pydantic - 数据验证和模型
pydantic==2.5.3
pydantic-settings==2.1.0

# ⚙️ python-dotenv - 环境变量管理
python-dotenv==1.0.0

# ⏰ dateparser - 时间解析处理
dateparser==1.2.0

# 🧪 pytest - 测试框架
pytest==7.4.4
pytest-asyncio==0.23.2
pytest-cov==4.1.0

# ================================
# 必要支持库 (Essential Dependencies)
# ================================

# Web框架支持
python-multipart==0.0.6  # FastAPI文件上传支持
starlette==0.35.1        # FastAPI底层框架

# 数据处理
numpy==1.26.2           # 数值计算
pandas==2.1.4           # 数据分析

# 异步和网络
httpx==0.26.0           # 异步HTTP客户端
aiofiles==23.2.1        # 异步文件操作

# 时间处理支持
python-dateutil==2.8.2  # 日期工具
pytz==2023.4            # 时区支持

# 日志和配置
structlog==23.2.0       # 结构化日志
click==8.1.7            # 命令行工具

# 开发工具
black==23.12.1          # 代码格式化
flake8==7.0.0           # 代码检查

# ================================
# 可选增强库 (Optional Enhancements)
# ================================

# 文本处理 (如需要中文支持)
# jieba==0.42.1

# 机器学习 (如需要NLP功能)
# scikit-learn==1.3.2
# spacy==3.7.2
# nltk==3.8.1

# 图数据可视化 (如需要图谱展示)
# networkx==3.2.1
# matplotlib==3.8.2

# Redis缓存 (如需要缓存功能)
# redis==5.0.1

# ================================
# 安装与配置说明
# ================================
#
# 🛠️ 安装步骤:
# 1. 创建虚拟环境: python -m venv venv
# 2. 激活环境:
#    Windows: venv\Scripts\activate
#    Linux/Mac: source venv/bin/activate  
# 3. 安装依赖: pip install -r requirements.txt
#
# ⚙️ 环境配置:
# 1. 复制 .env.example 为 .env
# 2. 配置以下环境变量:
#    - OPENAI_API_KEY=your_openai_api_key
#    - NEO4J_URI=bolt://localhost:7687
#    - NEO4J_USERNAME=neo4j
#    - NEO4J_PASSWORD=your_password
#
# 🚀 快速启动:
# - 运行主程序: python main.py
# - 启动API服务: uvicorn api.main:app --reload
# - 运行测试: pytest
#
# ================================ 